Add docs to emphasis shared-nothing assumption
